PopXL User Guide: add popxl.ops.var_updates to API section
Summary:
`popxl.ops.var_updates` is listed in the Supported Ops chapter but wasn't included in the API.

Some notes on the fixes. Adding `popxl.ops.var_updates` to the API caused many errors, mainly due to poorly formatted DocStrings. While fixing these I also fixed a bunch of other errors, for example renaming include files to `.inc` because these files are copied into the root doc directory when the doc is built and Sphinx treats them as files that are part of the main "index". Therefore Sphinx processes them twice, once because they are in the root directory and once when the files are actually included. This results in errors with links and Sphinx doesn't create the link because it has two definitions for a single anchor. Renaming them solves this.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #tech_docs, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha, shirazb

Maniphest Tasks: T70467

Differential Revision: https://phabricator.sourcevertex.net/D76822