Add conv transpose op to PopXL.
Summary: Added pybindings for the conv transpose op in a similar fashion to the other conv ops. More importantly, made the Conv transpose pattern mandatory becuase without it the conv transpose op cannot be lowered.

Test Plan: Added a unit test for adding the conv transpose op to the IR and that padding/output shape is computed according to the ONNX spec.

Reviewers: markk, matthewha, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: matthewha, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T69673

Differential Revision: https://phabricator.sourcevertex.net/D77394