Add conv_transpose_pow2scaled op to PopXL.
Summary:
Added the PopXL front-end for a ConvTranspose Op that also takes accepts a log2 scale tensor input. This function is essentially the same as `conv_pow2scaled` except for when the op is added to the Ir, the `ConvTransposed` op is addded instead of `Conv`.

Add a reinterpret in the grow method of the `ConvFlipWeightsOpx` because this was causing an unsigned char -> quarter copy error in poplar.

Depends on D77394

Test Plan:
Added tests for:
 - The `conv_transpose_pow2scaled` op. Note, most of these are added by parameterising the existing `conv_pow2scaled` tests by `conv_func` - an op that is either normal or transposed scaled convolution.
 - The conv transpose pattern
 - The `ConvOp::setup` method.

Reviewers: markk, matthewha, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: matthewha, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T69673

Differential Revision: https://phabricator.sourcevertex.net/D77609