Place PopArt assertion macros on the same line as end of Boost test macros in expression checking tests
Summary:
!ci_branch_mk2

It appears that on ubuntu 18, when the `POPART_CHECK_*` or `POPART_ASSERT_*` macros are inserted, the value of `__LINE__` from inside these macros is set to the end (closing bracket) of the `BOOST_CHECK_EXCEPTION` macro, whereas on ubuntu 20, it is set to the line on which the `POPART_*` macro is located. For this reason we keep the end of of the BOOST macros on the same line as the `POPART` macro with `clang-format on/off`, in the cases where clang-format spreads the macro arguments over multiple lines.

This fixes a test failure on ubuntu 20.04 debug CI: D77875

Test Plan: Failing tests were fixed in a Ubuntu 20.04 debug build container

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Maniphest Tasks: T71576

Differential Revision: https://phabricator.sourcevertex.net/D78154