Refactor Devicex to only create d2hWeightBuffers when the TensorData cannot be reused directly
Summary:
!ci_branch_mk2

---

`Devicex` currently creates a new buffer for the data of every weight to connect
to the device-to-host stream for that weight (which `weightsToHost` copies
across). These are the `d2hWeightBuffers: TensorId -> vec<char>` member.

This adds an extra O(N) memory, where N is the total size of all weights. For
large models, where these buffers are huge, this is unacceptable.

Instead, we re-use the existing `TensorData` buffers, connecting the streams to
them.

One caveat is that this is only valid when `opts.enableVariablesCaching` is
false. See `Devicex::needsIntermediaryD2hWeightBuffer` for the full explanation.

In PopXL, `opts.enableVariablesCaching` is always false, so that is not an
issue. The reason for this is that, in PopXL, you always have to construct your `Ir`,
passing all the tensor data, before creating the `Session`, which then does the
cache lookup. Therefore, there is no point serialising the data (in the previous run)
and then deserialising the data when we make the session, if the user is already
providing us the data each time.

The optimisation is also not valid when the variable retrieval mode of a
variable is `AllReplicas`, because then you need a buffer big enough to hold one
weight from every //replica//, not one for every group, which is what the
`TensorData` holds.

Therefore, in these two cases, we do create the intermediary `d2hWeightBuffer`.

---

To implement this, we abstract the `d2hWeightBuffers` member behind some new
helper functions, like `initD2hWeightBuffer()` and `getD2hWeightData()`. The
former only creates a buffer and inserts it into the `d2hWeightBuffers` map if
the conditions for the optimisation do not hold. The getter either returns a
pointer to the `TensorData`, or the `d2hWeightBuffer` if one was required.

---

As an additional change, we remove the unnecessary parameter
`isTensorDataSaved` from the helper function `checkCorrectionPopefMetadata`
in the test `executable_serialization_test.cpp`, as the correct value for this
parameter can always be wholly determined by `enableVariablesCaching`.

Test Plan:
Add a popxl integration test that inspects the logs to see if `d2hWeightBuffers`
were elided or not.

The real test we want to do is to inspect how the heap usage scales as we
increase the number of weights, but writing such a test (in a manner that would
not time out on CI) proved difficult. It could perhaps be made into an SDK test.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T70676, T70559

Differential Revision: https://phabricator.sourcevertex.net/D76962