Replace Devicex::h2dWeightBuffers with poplar::StreamCallbacks on the tensor data
Summary:
Depends on D76962
!ci_branch_mk2

---

To implement `weightsFromHost`, we create host-to-device host streams in Poplar
for each weight. We need to connect these streams to a buffer on host from which
they should copy. This is done with `pEngine->connectStream(stream, data)`.

When you do this, either the stream is `Broadcast`, and you pass one datum that
is then broadcast to each replica; or the stream is `Replicate` and you must
provide one datum for every replica.

When a variable is replica-grouped, the `TensorData` has one datum //per
group//. We therefore either must duplicate them into a new buffer that has one
datum per replica; or use:

```
pEngine->connectStreamToCallback(stream, replica, callback)
```
where `callback` copies `tdata[group(replica)]` to Poplar.

The former requires an extra `rf * datum_size` memory, whereas the latter is in-
place. In order to support large models where these buffers are so large that
the host memory is exhausted, we therefore choose the latter approach.

The former approach is the current approach in `Devicex`.

We therefore make two changes:

First, we remove the existing approach. That is, delete
`initialiseH2dWeightBuffers` and `deinitialiseH2dWeightBuffers`.

Secondly, we implement the new approach. That is, in
`loadEngineAndConnectStreams`, instead of using
`pEngine->connectStream(stream, h2dBuffer)`, we do
`pEngine->connectStreamToCallback(stream, replica, callback)`, where `callback`
is a new class `H2dWeightStreamCallback` that we implement, which copies
`tdata[group(replica)]` to Poplar.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T70676, T70557

Differential Revision: https://phabricator.sourcevertex.net/D77261