Don't infer modified regions by default for call op when input by ref
Summary:
!ci_branch_mk2

Inferring the regions for tensor by ref inputs of a call op were taking in total 30mins for GPT3. After discussing it with Oskar we agreed that it is easier to disable it by default for all models as it doesn't add much value.

The linter also made some changes. Line 135 of call.py op is the only code change

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Subscribers: shirazb, matthewha

Maniphest Tasks: T70572

Differential Revision: https://phabricator.sourcevertex.net/D78475