Update readme
Summary:
!ci_branch_mk2
First, on ubuntu 20.04 installs, the following packages have versions that are compatible with PopART, and so do not require install steps:
- CMake (3.16.3)
- Python (3.8)
- libprotobuf-dev (3.6.1.3)
The install steps for these packages are removed from the readme.
Second, the README had a long list of pip installations. These are covered by the `requirements.txt`s, so the README is now refers to these instead.
Third, the section referring to graphcore examples had a broken link, which is now correct.
Fourth, the the docker file don't exist, so the section referring to these is removed.
Last, some rephrasing and typos to be more concise.

Test Plan: I read it.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: jamiep, shirazb

Maniphest Tasks: T72102

Differential Revision: https://phabricator.sourcevertex.net/D78843