python version of host scale calculator
Summary:
!ci_branch_mk2

Python version to calculate the fp8 log2_scale with mse metric on host. It takes the middle scaling bias when multiple bias have the same mse. The result is different from the wrapper of poplar that takes the first bias that has the min mse and no negate.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, simonl, #popart, markk

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #popart, markk

Subscribers: markk, awf, sergiop

JIRA Issues: FEA-30

Differential Revision: https://phabricator.sourcevertex.net/D80681