Add RUN_SERIAL TRUE to multicard tests
Summary:
When the Tyan machine is removed from CI then all tests will be run on the POD.
To speed things up the tests will be run in parallel without the multicard label.
So this patch forces tests that want to use lots of IPUs to run serially.

!ci_branch_mk2

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

JIRA Issues: PCII-104

Differential Revision: https://phabricator.sourcevertex.net/D82984