preload popef file if env var is set
Summary:
!ci_branch_mk2

Adds a sequential read of the popef file in Reader::checkFileForValidPoplarExecutable() before the Reader constructor. The preload will only be executed if a corresponding environment variable is set, and so default Popart behaviour is unchanged.

The reason for adding this preload is because of intermittent issues we've observed on some platforms where the first read of a file is slow, and the Reader constructor is particularly affected, so that for a user, loading a precompile popef executable can take 20 minutes or more (sometimes longer than compiling the model from scratch). The preload will read the file much more quickly in these cases, and after the first read of the file, subsequent read speeds are expected to go up by ~100x, and so the Reader constructor is then no longer slow.

Test Plan:
Ran the stablediffusion2 notebook on a custom Popart build containing these changes. Default behaviour is unchanged. When the environment variable POPART_PRELOAD_POPEF is set to `full-preload`, we observe in the Popart logs (level=DEBUG) that the preload has been performed, and the user observes much faster model compilation in the UI.

Also tested with `pytest popart/tests/integration/popxl/test_cached_executables.py::test_popart_preload`

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, alexandrep, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Subscribers: shirazb, grahamh

JIRA Issues: AIP-147

Differential Revision: https://phabricator.sourcevertex.net/D83495