Add more efficient way of calculating copy costs in outliner
Summary:
!ci_branch_mk2

This change is an experimental change to improve the performance of the SubgraphOutline transform in PopART following issues raised in T74421 (note: outlining isn't the only problematic area in terms of PopART performance for this model).

This change follows an observation that on a smaller model `m.onnx` in T74362 that is similar in nature to the much larger model in T74421 that a large proportion of time is spent in `IoSubgraphCostModel::value` itself. We assume this significantly contributes to SubgraphOutline's time.

We speculate that most of the time in `IoSubgraphCostModel::value` is spent calculating [a heuristic for] copying costs. In this diff we try and pre-compute as much of that copying cost in the first call to  `IoSubgraphCostModel::value` as possible.

Time breakdown of `m.onnx` without this change:
```
Ir preparation complete. Breakdown of compile time so far:
Scope                                      Time [s]             Count  Percentage
-----                                      --------             -----  ----------
Growing partial AliasModel                 1029.018034          51972        40 %
**SubgraphOutline                            691.027348           25752        27 %**
[Scheduler] defaultAnnotate                451.522207            1268        17 %
OverlapIO                                  76.123301              359         3 %
Scheduling. TCOs=0 Refine=0 CacheHit=Cold  64.362276              112         2 %
Tensor::anyAlias                           54.668109           104191         2 %
Scheduling. TCOs=0 Refine=0 CacheHit=Hot   42.295321              145         2 %
MergeExchange                              34.720678                7         1 %
MergeDuplicateOps                          33.519943                1         1 %
Total                                      2586.445583            n/a       100 %
Unaccounted for                            0.139842               n/a         0 %
Accounted for                              2586.305741            n/a       100 %
```

Time breakdown of `m.onnx` with this change:
```
Ir preparation complete. Breakdown of compile time so far:
Scope                                      Time [s]             Count  Percentage
-----                                      --------             -----  ----------
Growing partial AliasModel                 1078.734242          52514        51 %
[Scheduler] defaultAnnotate                517.285786            1277        25 %
OverlapIO                                  77.036072              359         4 %
**SubgraphOutline                            71.443824            25786         3 %**
Scheduling. TCOs=0 Refine=0 CacheHit=Cold  65.532556              110         3 %
Tensor::anyAlias                           57.184291           105305         3 %
Scheduling. TCOs=0 Refine=0 CacheHit=Hot   43.335668              143         2 %
MergeExchange                              36.002367                7         2 %
MergeDuplicateOps                          35.237735                1         2 %
Tensor::getVirtualGraphIdAndTileSetUnsafe  25.342620           168971         1 %
Inplace::apply                             23.328948             4862         1 %
Total                                      2098.569574            n/a       100 %
Unaccounted for                            0.142319               n/a         0 %
Accounted for                              2098.427255            n/a       100 %
```

It is unknown whether these results translate to improved Outlining for T74421.

Test Plan: Write a unit test for `IoSubgraphCostModel::value` for the existing code, checking the values returned for some model. Check that the new code returns the same values (modulo float precision) as `master` does (see test in diff).

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha, leiw

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha, leiw

Subscribers: hanzhao, leiw, matthewha

Maniphest Tasks: T74421

Differential Revision: https://phabricator.sourcevertex.net/D84506